Add Test, Run and Requirement entities
The plugin now fetch WorkItem, Test, Run and Requirement entities.
All the entities are filtered by the authenticated user and specific criteria such as phase or status.
Most of the metadata about the entities is now defined and stored in MyWorkMetadata class, only the fetch criteria is still defined in OctaneServices class.